+++ src/main/java/com/mystchonky/arsocultas/datagen/recipe/EnchantingAppProvider.java
@@ -2,6 +2,8 @@ package com.mystchonky.arsocultas.datagen.recipe;
 
 import com.hollingsworth.arsnouveau.api.enchanting_apparatus.EnchantingApparatusRecipe;
 import com.hollingsworth.arsnouveau.common.datagen.ApparatusRecipeProvider;
+import com.hollingsworth.arsnouveau.setup.registry.ItemsRegistry;
+import com.klikli_dev.occultism.registry.OccultismItems;
 import com.mystchonky.arsocultas.datagen.DataProvider;
 import net.minecraft.data.CachedOutput;
 import net.minecraft.data.DataGenerator;
@@ -17,15 +19,15 @@ public class EnchantingAppProvider extends ApparatusRecipeProvider {
     @Override
     public void collectJsons(CachedOutput pOutput) {
         //example of an apparatus recipe
-        /*
         recipes.add(builder()
                 .withReagent(ItemsRegistry.SOURCE_GEM)
-                .withPedestalItem(4, Recipes.SOURCE_GEM)
-                .withResult(ItemsRegistry.BUCKET_OF_SOURCE)
-                .withSource(100)
+                .withPedestalItem(1, ItemsRegistry.CONJURATION_ESSENCE)
+                .withPedestalItem(1, ItemsRegistry.MANIPULATION_ESSENCE)
+                .withPedestalItem(1, OccultismItems.DATURA)
+                .withResult(OccultismItems.SPIRIT_ATTUNED_GEM)
+                .withSourceCost(500)
                 .build()
         );
-         */
 
         for (EnchantingApparatusRecipe g : recipes) {
             if (g != null) {
@@ -42,6 +44,6 @@ public class EnchantingAppProvider extends ApparatusRecipeProvider {
 
     @Override
     public String getName() {
-        return "Example Apparatus";
+        return "Ars Ocultas - Enchanting Apparatus";
     }
 }